Missed issues in last commit - Didn't actually implement `First()`, oops. - Needed to pull out scan constraint construction from query ranges - Bad cast in manifest emitter, causing warnings about downcasting

+++ src/emit/ShaderManifestEmitter.cpp
@@ -695,7 +695,7 @@ namespace
         {
             std::string_view valName = axis.ValueDomain == AxisValueDomain::Type ? value.AsType(axis) : value.AsEnumCase(axis);
             uint32_t valNameIdx = strings.Add(valName);
-            tables.Values.emplace_back(static_cast<int64_t>(valNameIdx));
+            tables.Values.emplace_back(valNameIdx);
         }
     }
